Refactor root structure (#494)
* moved index.jsx to index.tsx

- now it's typesafe :sunglasses:
- renamed the script import in index.html to match

* moved App to the root, next to index.tsx

* moved the theme to its own custom provider

* converted App to a function component

* moved react-app-env.d.ts to vite-env.d.ts

* added query provider

* simplified index.tsx by moving theme to provider

* forgot to add snackbar provider
