+++ index.html
@@ -109,6 +109,6 @@
     <body style="overflow: hidden">
         <noscript> You need to enable JavaScript to run this app. </noscript>
         <div id="root" style="height: 100vh; position: fixed; top: 0; width: 100%"></div>
-        <script type="module" src="/src/index.jsx"></script>
+        <script type="module" src="/src/index.tsx"></script>
     </body>
 </html>
