+++ src/vite-env.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />
+
+declare module '*.svg'
